move to utils.go
